requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-extensions==3.0.4
django-templated-mail==1.1.1
djangorestframework==3.11.1
djoser==2.0.3
gunicorn==20.0.4
idna==2.10
numpy==1.19.1
pandas==1.1.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
